tetragon: Move rawSyscallEnter to raw_syscall_enter_linux.go
splitting encoder package into platform independent and linux specific listings
